Move storage declaration to ty module.
