+++ sway-core/src/semantic_analysis/program.rs
@@ -1,7 +1,4 @@
-use super::{
-    storage_only_types, TyAstNode, TyAstNodeContent, TyFunctionDeclaration, TyImplTrait,
-    TyStorageDeclaration,
-};
+use super::{storage_only_types, TyAstNode, TyAstNodeContent, TyFunctionDeclaration, TyImplTrait};
 use crate::{
     declaration_engine::declaration_engine::{de_get_function, de_get_impl_trait, de_get_storage},
     error::*,
@@ -158,7 +155,7 @@ impl TyProgram {
                 .find(|decl| matches!(decl, ty::TyDeclaration::StorageDeclaration(_)));
 
             if let Some(ty::TyDeclaration::StorageDeclaration(decl_id)) = storage_decl {
-                let TyStorageDeclaration { span, .. } = check!(
+                let ty::TyStorageDeclaration { span, .. } = check!(
                     CompileResult::from(de_get_storage(decl_id.clone(), &decl_id.span())),
                     return err(warnings, errors),
                     warnings,
